requirements.txt
pytest==8.2.2
requests==2.32.3
pydantic==2.8.2
allure-pytest==2.13.5
python-dotenv==1.0.1